+++ public/atompagebuilder/app/_app/immutable/chunks/5-df124c8f.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/pages/(editor)/editor/_page.svelte-30b8adaf.js";export{t as component};